\name{exampleMethyGenoSet}
\alias{exampleMethyGenoSet}
\docType{data}
\title{
Example MethyGenoSet dataset
}
\description{
Example MethyGenoSet dataset, which includes eight randomly picked cancer cell line samples from two tissue types. To save space, only 21 chromosome data was included.
}
\usage{data(exampleMethyGenoSet)}
\details{
Example MethyGenoSet dataset, which includes eight randomly picked cancer cell line samples from two tissue types. To save space, only 21 chromosome data was included.
}

\examples{
data(exampleMethyGenoSet)
class(exampleMethyGenoSet)
colnames(exampleMethyGenoSet)
exampleMethyGenoSet
}
\keyword{datasets}
